(Daniel) recalibrated second arm, got some logic for two arm case implemented.
We will have to retune the angle for placing, it's really causing some issues here. The second value helps to go 'side by side' and the third jonit is further up... if that makes sense.

And the logic isn't entirely correct when we have uneven amounts of blocks to do, per arm